backend/scheduler/requirements.txt
apscheduler
psycopg2
psycopg2-binary
yfinance
python-dotenv